Change the entiere architecture
I choosed to use Rack to improve and delegate the parsing of the user
request. The aim is to simplify my own code. It has a cost for unitary
tests, which are more difficults to setup, at least for the moment.

It was a requierement to improve the archaic architecture I made in the
previous versions. I also improved the system's logic by more
instanciation, more smartly the differents components of the system

- improve bin/server with more verbose on --debug
- Simplification of controller api and creation
- Drop the routing management for trailing / and get it from
  Rack::Request. Then, the router has a \/* rule at the end of the
  regexp match
- Simplify the render helpers from the Router
- Simplify the execute method from the Router by using
- Rack::Request (errors and controllers are the only problematics of the
  Router)
- Change Router and the Responder to Class from Module
- Instanciate a Router in the Server
- Instanciate a Responder in Router
- Instanciate the Controller in the Responder (no more in the Router)

I also updated the unitary test (cleaning some of them in the same
time), but there is a real work to do on the documentation and unitary
tests.
I'll have to investigate for checking the unitary test coverage, and
also create complete feature tests.