Improve routing helpers
In the routing helpers, you can specifie the url out out the hash. It
mean that the helpers take one more argument, before the hash, which is
a string, the url to use. The url has no more to be specified in the
hash.

Note: if the hash already contains a url, it will not be erased by the
previous parameter